pyproject.toml
[project]
name = "reddit-health"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.0.1",
    "datasets>=3.1.0",
    "huggingface-hub>=0.36.0",
    "ipykernel>=6.29.5",
    "nmslib>=2.1.2",
    "numpy>=1.26.0,<2.0.0",
    "openai>=2.2.0",
    "pandas>=2.0.3",
    "scispacy>=0.5.5",
    "spacy>=3.7.5",
    "tqdm>=4.67.1",
    "transformers>=4.46.3",
    "scikit-learn==1.1.3",
    "en-core-sci-lg",
    "torch>=2.5.1",
    "faiss-gpu>=1.7.2",
    "vllm>=0.7.3",
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }
en-core-sci-lg = { url = "https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_lg-0.5.4.tar.gz" }

[tool.uv]
index-strategy = "unsafe-best-match"
